Adicionada autenticação as rotas

+++ routes/api.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\CategoriaController;
 use App\Http\Controllers\VideoController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -14,21 +15,28 @@ use Illuminate\Support\Facades\Route;
 | is assigned the "api" middleware group. Enjoy building your API!
 |
 */
-Route::get('/categorias', [CategoriaController::class, "index"]);
-Route::get('/categorias/{id}', [CategoriaController::class, "show"]);
-Route::get('/categorias/{id}/videos', [CategoriaController::class, "videos"]);
-Route::post('/categorias', [CategoriaController::class, "store"]);
-Route::put('/categorias/{id}', [CategoriaController::class, "update"]);
-Route::delete('/categorias/{id}', [CategoriaController::class, "destroy"]);
+Route::middleware('auth:api')->get('/categorias', [CategoriaController::class, "index"]);
+Route::middleware('auth:api')->get('/categorias/{id}', [CategoriaController::class, "show"]);
+Route::middleware('auth:api')->get('/categorias/{id}/videos', [CategoriaController::class, "videos"]);
+Route::middleware('auth:api')->post('/categorias', [CategoriaController::class, "store"]);
+Route::middleware('auth:api')->put('/categorias/{id}', [CategoriaController::class, "update"]);
+Route::middleware('auth:api')->delete('/categorias/{id}', [CategoriaController::class, "destroy"]);
 
-Route::get('/videos', [VideoController::class, "index"]);
-Route::get('/videos/{id}', [VideoController::class, "show"]);
-Route::post('/videos', [VideoController::class, "store"]);
-Route::put('/videos/{id}', [VideoController::class, "update"]);
-Route::delete('/videos/{id}', [VideoController::class, "destroy"]);
+Route::middleware('auth:api')->get('/videos', [VideoController::class, "index"]);
+Route::middleware('auth:api')->get('/videos/{id}', [VideoController::class, "show"]);
+Route::middleware('auth:api')->post('/videos', [VideoController::class, "store"]);
+Route::middleware('auth:api')->put('/videos/{id}', [VideoController::class, "update"]);
+Route::middleware('auth:api')->delete('/videos/{id}', [VideoController::class, "destroy"]);
+
+Route::middleware('auth:api')->post('/usuarios', [UserController::class, "store"]);
+Route::middleware('auth:api')->put('/usuarios/alterar-senha/{id}', [UserController::class, "updatePassword"]);
+Route::middleware('auth:api')->put('/usuarios/alterar-token/{id}/', [UserController::class, "updateToken"]);
+Route::middleware('auth:api')->put('/usuarios/{id}', [UserController::class, "update"]);
+Route::middleware('auth:api')->delete('/usuarios/{id}', [UserController::class, "destroy"]);
+
 
 /*
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
+Route::middleware('auth:api')->middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 */
